Telegram-bot/requirements.txt.txt
python-telegram-bot==20.7
deep-translator==1.11.4